cloud_reid_api/requirements.txt
aiohttp==3.6.0
async-timeout==3.0.1
attrs==19.1.0
bcrypt==3.1.7
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.29.1
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-datastore==1.8.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
h11==0.8.1
httptools==0.0.13
idna==2.8
multidict==4.5.2
passlib==1.7.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pydantic==0.26
PyJWT==1.7.1
python-multipart==0.0.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
starlette==0.12.0
urllib3==1.25.3
uvicorn==0.7.2
uvloop==0.12.2
websockets==7.0
yarl==1.3.0
